updated: layout.tsx, dataTableSearchField.tsx, and dataTableToolbar.tsx; removed quickFilter.tsx and moved its content to dataTable/quickFilter.tsx. Updated styles
