+++ src/app/(private)/layout.tsx
@@ -9,7 +9,7 @@ export default function RootLayout({
       <Navbar />
       <div className="grid w-full md:grid-cols-[220px_1fr] lg:grid-cols-[280px_1fr] h-[calc(100dvh-3.5rem)]">
         <Sidebar />
-        <main className="flex flex-1 flex-col gap-4 p-4 lg:gap-6 lg:p-6 overflow-auto">
+        <main className="flex flex-1 flex-col gap-4 p-4 lg:gap-6 lg:p-6 overflow-y-scroll">
           {children}
         </main>
       </div>
